pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "retitle"
version = "0.1.0"
description = "Rename messy media filenames using TV show/movie metadata"
requires-python = ">=3.11"
dependencies = [
    "guessit>=3.8",
    "click>=8.0",
    "requests>=2.28",
    "python-dotenv>=1.0",
]

[project.scripts]
retitle = "retitle.cli:cli"

[project.gui-scripts]
retitle-gui = "retitle.gui:main"
